Separate Library/GridSurface completed
